// packages/frontend/src/generated/types/DispatchResponse.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { BunkerType } from './BunkerType.ts'
import type { DispatchMethod } from './DispatchMethod.ts'
import type { DispatchPurpose } from './DispatchPurpose.ts'
import type { DocumentStatus } from './DocumentStatus.ts'

/**
 * @description Response DTO for the `dispatch_document` entity.
 */
export interface DispatchResponse {
  bunkerType?: null | BunkerType
  /**
   * @type string, uuid
   */
  contractorId: string
  /**
   * @type string
   */
  createdAt: string
  /**
   * @type string, uuid
   */
  createdBy: string
  /**
   * @type string
   */
  date: string
  /**
   * @type string,null
   */
  deletedAt?: string | null
  /**
   * @type string,null, uuid
   */
  deletedBy?: string | null
  /**
   * @type string,null, uuid
   */
  destinationBaseId?: string | null
  /**
   * @type string
   */
  dispatchMethod: DispatchMethod
  /**
   * @type string
   */
  dispatchPurpose: DispatchPurpose
  /**
   * @type string
   */
  documentNumber: string
  /**
   * @type string,null
   */
  endCargoOps?: string | null
  /**
   * @type string,null
   */
  executedAt?: string | null
  /**
   * @type string,null, uuid
   */
  executedBy?: string | null
  /**
   * @type string,null, uuid
   */
  exporterId?: string | null
  /**
   * @type string, uuid
   */
  id: string
  /**
   * @type string, uuid
   */
  originDbId: string
  /**
   * @type string,null, uuid
   */
  portId?: string | null
  /**
   * @type string,null
   */
  receiverEntity?: string | null
  /**
   * @type string,null
   */
  revertedAt?: string | null
  /**
   * @type string,null, uuid
   */
  revertedBy?: string | null
  /**
   * @type string,null
   */
  startCargoOps?: string | null
  /**
   * @type string
   */
  status: DocumentStatus
  /**
   * @type string
   */
  updatedAt: string
  /**
   * @type string, uuid
   */
  updatedBy: string
}